requirements-dev.txt
check-manifest
dask
numpydoc
pandas>=2
pytest
scipy
setuptools_scm
sphinx
sphinx_rtd_theme
twine
xarray